[bitset] Rename bitset module
