+++ src/pdata/mod.rs
@@ -1,6 +1,6 @@
 pub mod array;
 pub mod array_walker;
-pub mod bitset_walker;
+pub mod bitset;
 pub mod btree;
 pub mod btree_builder;
 pub mod btree_merge;
